Introduce a general Shape data type
